% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_interfaces.R
\name{generic_C_interface}
\alias{generic_C_interface}
\title{Generic C interface}
\usage{
generic_C_interface(x, C_fct, ...)
}
\arguments{
\item{x}{a numeric \code{"uts"} object with finite, non-NA observation values.}

\item{C_fct}{the name of the C function to call.}

\item{\dots}{further arguments passed to the C function.}
}
\description{
Generic interface for C-functions with inputs (values, times, length(values), values_new, ...) and output (values_new). Example: sma, rollingMax, ema, ...
}
\examples{
# SMA_last
generic_C_interface(ex_uts(), "sma_last", width=ddays(1))

# One- vs. two-sided window
generic_C_interface(ex_uts(), "rolling_num_obs", width_before=dhours(6), width_after=dhours(0))
generic_C_interface(ex_uts(), "rolling_num_obs", width_before=dhours(6), width_after=dhours(6))
}
\keyword{internal}

