% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rolling_apply_specialized.R
\name{have_rolling_apply_specialized}
\alias{have_rolling_apply_specialized}
\title{Specialized Rolling Apply Available?}
\usage{
have_rolling_apply_specialized(x, FUN, by = NULL)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{FUN}{see \code{\link{rolling_apply_specialized}}.}

\item{by}{see \code{\link{rolling_apply_specialized}}.}
}
\description{
Check whether \code{\link{rolling_apply_specialized.uts}} can be called for a given \code{\link{uts}} object with arguments \code{FUN} and \code{by}.
}
\examples{
have_rolling_apply_specialized(ex_uts(), FUN=mean)
have_rolling_apply_specialized(ex_uts(), FUN="mean")
have_rolling_apply_specialized(ex_uts(), FUN=mean, by=ddays(1))
have_rolling_apply_specialized(uts(NA, Sys.time()), FUN=mean)

FUN <- mean
have_rolling_apply_specialized(ex_uts(), FUN=FUN)
}
\keyword{internal}

