% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper.R
\name{rev.uts}
\alias{rev.uts}
\title{Reverse Observations}
\usage{
\method{rev}{uts}(x)
}
\arguments{
\item{x}{a \code{"uts"} object.}
}
\description{
Reverse the observation values and times inside the time window bounded by the first and last observation time.
}
\examples{
rev(ex_uts())

# Reversing a "uts" reverses the vector of observation values
ex_uts()$values
rev(ex_uts())$values

# Reversing a "uts" reverses the vector of observation time differences
diff(ex_uts()$times)
diff(rev(ex_uts())$times)
}
\keyword{internal}

