% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rolling_apply.R
\name{rolling_apply}
\alias{rolling_apply}
\title{Apply Rolling Function}
\usage{
rolling_apply(x, ...)
}
\arguments{
\item{x}{a time series object.}

\item{width}{a non-negative \code{\link[lubridate]{duration}} object, specifying the temporal width of the rolling time window.}

\item{FUN}{a function to be applied to the vector of observation values within the rolling time window.}

\item{\dots}{arguments passed to \code{FUN}.}

\item{by}{a positive \code{\link[lubridate]{duration}} object. Calculate \code{FUN} on a sequence of time points with this spacing, rather than at every observation time of \code{x}.}

\item{align}{}
}
\description{
Apply a function to the time series values in a rolling time window.
}

