% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rolling_apply.R
\name{rolling_apply.uts}
\alias{rolling_apply.uts}
\title{}
\usage{
\method{rolling_apply}{uts}(x, width, FUN, ..., by = NULL, align = "right")
}
\description{

}
\examples{
rolling_apply(ex_uts(), ddays(0.1), FUN="mean", by=ddays(0.1))
plot(rolling_apply(SPX, ddays(90), FUN="mean"))
plot(rolling_apply(SPX, ddays(90), FUN="max", by=ddays(90)))
plot(rolling_apply(SPX, ddays(90), FUN="min"))
system.time(window_apply(SPX, ddays(90), FUN="mean"))   # 0.56s
}

