% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rolling_apply.R
\name{rolling_apply_helper}
\alias{rolling_apply_helper}
\title{Apply Rolling Function (Static Version)}
\usage{
rolling_apply_helper(x, left, end, FUN, ..., align = "right")
}
\arguments{
\item{x}{a time series object.}

\item{end}{a strictly increasing \code{\link{POSIXct}} object of same length as \code{start}, and with \code{start[i] <= end[i]} for each \code{1 <= i <= length(start)}. Specifies the end times of the time windows.}

\item{FUN}{a function to be applied to the vector of observation values in each close time interval \code{[start[i], end[i]]}.}

\item{align}{either \code{"right"} (the default), \code{"left"}, or \code{"center"}. Specifies the position of each output time inside the corresponding time window.}

\item{start}{a strictly increasing \code{\link{POSIXct}}, specifying the start times of the time windows.}

\item{\dots}{arguments passed to \code{FUN}.}
}
\description{
Apply a function to the time series values in a squence of user-defined time windows.
}
\examples{
start <- seq(as.POSIXct("2007-11-08"), as.POSIXct("2007-11-09 12:00:00"), by="12 hours")
end <- start + dhours(4)
rolling_apply_helper(ex_uts(), start, end, FUN=mean)
rolling_apply_helper(ex_uts(), start, end, FUN=mean, align="left")
rolling_apply_helper(ex_uts(), start, end, FUN=mean, align="center")
}
\keyword{internal}

