% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rolling_apply_specialized.R
\name{rolling_apply_specialized.uts}
\alias{rolling_apply_specialized.uts}
\title{Rolling Average/Length/Minimun/Maximum/Sum}
\usage{
\method{rolling_apply_specialized}{uts}(x, width, FUN)
}
\description{
Rolling Average/Length/Minimun/Maximum/Sum
}
\examples{
rolling_apply_specialized(ex_uts(), ddays(1), FUN=sum)
rolling_apply_specialized(ex_uts(), ddays(1), FUN="sum")

rolling_apply_specialized(ex_uts(), ddays(1), FUN=sum) - rolling_apply(ex_uts(), ddays(1), FUN=sum)
}

