% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rolling_specialized.R
\name{rolling_sum.uts}
\alias{rolling_sum.uts}
\title{Rolling Sum}
\usage{
rolling_sum.uts(x, width)
}
\arguments{
\item{x}{a numeric time series object.}

\item{width}{a finite, positive \code{\link[lubridate]{duration}} object, specifying the temporal width of the rolling time window.}
}
\description{
Calculate the rolling sum of all observation values in a rolling time window. This method is a fast, specialized implementation of \code{\link{rolling_apply}} with \code{FUN=sum} and \code{by=NULL} (i.e. move the rolling time window one observation at a time, rather than by a fixed temporal amount).
}
\examples{
rolling_sum.uts(ex_uts(), ddays(1))
rolling_sum.uts(ex_uts(), ddays(1)) - rolling_apply(ex_uts(), ddays(1), FUN=sum)
}
\references{
Eckner, A. (2010) \emph{Algorithms for Unevenly Spaced Time Series: Moving Averages and Other Rolling Operators}.
}
\seealso{
\code{\link{rolling_apply}}
}
\keyword{internal}

