% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rolling_apply.R
\name{rolling_time_window}
\alias{rolling_time_window}
\title{Rolling Time Window}
\usage{
rolling_time_window(start, end, width, by, align = "right")
}
\arguments{
\item{start}{a \code{\link{POSIXct}} object or coercible using \code{\link{as.POSIXct}}. The alignment point of the first time window.}

\item{end}{a \code{\link{POSIXct}} object or coercible using \code{\link{as.POSIXct}}. The latest possible alignment point.}

\item{width}{a non-negative \code{\link[lubridate]{duration}} object, specifying the temporal width of the rolling time window.}

\item{by}{a positive \code{\link[lubridate]{duration}} object. The temporal spacing between alignment points (and therefore also start times and end times) of adjacent time windows.}

\item{align}{either \code{"right"} (the default), \code{"left"}, or \code{"center"}. Specifies how the rolling windows should be aligned relative to the sequence of time points generated between \code{start} and \code{end}.}
}
\value{
A list with two \code{POSIXct} objects of equal length, specifying the start and end times of the rolling time window.
}
\description{
Generate a sequence of start times and end times for a rolling time window of specified width.
}
\examples{
rolling_time_window(start="2015-01-01", end="2015-06-30", width=ddays(90), by=ddays(30))
}
\keyword{internal}

