% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rolling_apply.R
\name{rolling_time_window_indices}
\alias{rolling_time_window_indices}
\title{Rolling Time Window Indices}
\usage{
rolling_time_window_indices(times, start_times, end_times)
}
\arguments{
\item{times}{a \code{\link{POSIXct}} object of strictly increasing time points.}

\item{start_times}{a strictly increasing \code{\link{POSIXct}} object, specifying the start times of the time windows.}

\item{end_times}{a strictly increasing \code{\link{POSIXct}} object of same length as \code{start}, and with \code{start[i] <= end[i]} for each \code{1 <= i <= length(start)}. Specifies the end times of the time windows.}
}
\value{
A list with two integer vectors of equal length, specifying the start and end index in \code{times} of each rolling time window.
}
\description{
For a sorted sequence of time points, determine the start and end indices inside a rolling time window.
}
\examples{
tmp <- rolling_time_window(start="2015-01-01", end="2015-06-30", width=ddays(90), by=ddays(30))
times <- seq(as.POSIXct("2014-12-01"), as.POSIXct("2015-12-30"), by="week")
rolling_time_window_indices(times, tmp$start_times, tmp$end_times)
}
\keyword{internal}

