% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sma.R
\name{sma}
\alias{sma}
\alias{sma.uts}
\title{Simple Moving Average (SMA)}
\usage{
sma(x, ...)

\method{sma}{uts}(x, tau, type = "last", ...)
}
\arguments{
\item{x}{a \code{"uts"} object with \code{\link{logical}} observation values.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Calculate a simple moving average (SMA) of a time series.
}
\section{Methods (by class)}{
\itemize{
\item \code{uts}: simple moving average for \code{"uts"} objects.
}}
\examples{
sma(ex_uts(), ddays(1))
}
\seealso{
\code{\link{ema}} for exponential moving averages.
}

