% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sma.R
\name{sma_equal_R}
\alias{sma_equal_R}
\title{R implementation of sma(..., type="equal")}
\usage{
sma_equal_R(x, tau)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{tau}{a non-negative \code{\link[lubridate]{duration}} object, specifying the temporal width of the rolling time window.}
}
\description{
This function exists solely for testing the C implementation.
}
\examples{
sma_equal_R(ex_uts(), ddays(1)) - sma(ex_uts(), ddays(1), type="equal")
}
\keyword{internal}

