% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma.R
\name{sma_linear_R}
\alias{sma_linear_R}
\title{R implementation of sma(..., interpolation="linear")}
\usage{
sma_linear_R(x, width)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{width}{a positive \code{\link[lubridate]{duration}} object, specifying the temporal width of the rolling time window.}
}
\description{
This function is identical to \code{\link{sma}} with \code{interpolation="linear"}. It exists solely for testing the C implementation.
}
\examples{
sma_linear_R(ex_uts(), ddays(1)) - sma(ex_uts(), ddays(1), interpolation="linear")
}
\keyword{internal}

