% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{utsOperators-internal}
\alias{utsOperators-internal}
\title{Internal Functions}
\description{
The internal functions listed below might be of interest to developers seeking to extend the package.
}
\details{
C interfaces:
\itemize{
  \item \code{\link{generic_C_interface}}
  \item \code{\link{generic_C_interface_rolling}}
}

\code{uts} methods:
\itemize{
  %\item \code{\link{ema_equal_R}}
  %\item \code{\link{ema_last_R}}
  %\item \code{\link{ema_next_R}}
  %\item \code{\link{ema_linear_R}
  \item \code{\link{sma_equal_R}}
  %\item \code{\link{sma_last_R}}
  %\item \code{\link{sma_next_R}}
  %\item \code{\link{sma_linear_R}}
  \item \code{\link{rev}}
}
}

