% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ema.R
\name{ema}
\alias{ema}
\title{Exponential Moving Average (SMA)}
\usage{
ema(x, ...)
}
\arguments{
\item{x}{a time series object.}

\item{tau}{a \code{\link[lubridate]{duration}} object, specifying the effective temporal length of the EMA}

\item{type}{the type of the EMA. Either \code{"equal"}, \code{"last"}, \code{"next"}, or \code{"linear"}. See below for details}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Calculate an exponential moving average (EMA) of a time series.
}
\details{
Currently, four different EMAs types are supported for \code{"uts"} objects. Each type puts different weights on past observation values. The: \itemize{
\item \code{equal}:
\item \code{last}:
\item \code{next}:
\item \code{linear}:
}
See the reference below for details for precise definitions and on why one would use one EMA type over another.
}
\references{
Eckner, A. (2010) \emph{Algorithms for Unevenly Spaced Time Series: Moving Averages and Other Rolling Operators}.
}
\seealso{
\code{\link{sma}} for simple moving averages.
}

