% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/C_interfaces.R
\name{generic_C_interface}
\alias{generic_C_interface}
\title{Generic C interface}
\usage{
generic_C_interface(x, C_fct, NA_method = "ignore", ...)
}
\arguments{
\item{x}{a numeric \code{"uts"} object.}

\item{C_fct}{the name of the C function to call.}

\item{NA_method}{the method for dealing with \code{NA}s. Either \code{"fail"}, \code{"ignore"}, \code{"omit"}.}

\item{\dots}{further arguments passed to the C function.}
}
\description{
Generic interface for C-functions with inputs (values, times, length(values), values_new, ...) and output (values_new). Example: sma, rollingMax, ema, ...
}
\examples{
# Prepare sample 'uts"
x <- ex_uts()
x$values[2] <- NA

# SMA_last
generic_C_interface(x, "sma_last", width=ddays(1))
generic_C_interface(x, "sma_last", width=ddays(1), NA_method="omit")
}
\keyword{internal}

