% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/C_interfaces.R
\name{generic_C_interface_rolling}
\alias{generic_C_interface_rolling}
\title{Generic C interface for rolling time series operators}
\usage{
generic_C_interface_rolling(x, tau, NA_method = "reinsert", C_fct, ...)
}
\arguments{
\item{tau}{a \code{\link[lubridate]{duration}} object, specifying the temporal length of the rolling time window.}

\item{\dots}{further arguments passed to \code{\link{generic_C_interface}}.}
}
\description{
Generic interface for rolling operators with one parameter tau (window width), and zero or more other parameters
}
\keyword{internal}

