% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/C_interfaces.R
\name{generic_C_interface_rolling}
\alias{generic_C_interface_rolling}
\title{Generic C interface for rolling time series operators}
\usage{
generic_C_interface_rolling(x, tau, ...)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{tau}{a \code{\link[lubridate]{duration}} object, specifying the temporal length of the rolling time window.}

\item{\dots}{further arguments passed to \code{\link{generic_C_interface}}.}
}
\description{
This function is a convenience wrapper around \code{\link{generic_C_interface}} that adds argument checking of the rolling window width \code{tau}.
}
\examples{
# Prepare sample 'uts"
x <- ex_uts()
x$values[2] <- NA

# SMA_eqqual
generic_C_interface_rolling(x, tau=ddays(1), C_fct="sma_equal")
generic_C_interface_rolling(x, tau=ddays(1), C_fct="sma_equal", NA_method="omit")

# EMA_last
generic_C_interface_rolling(x, tau=ddays(1), C_fct="ema_equal")
generic_C_interface_rolling(x, tau=ddays(1), C_fct="ema_equal", NA_method="omit")

# Forward-looking SMA
generic_C_interface_rolling(x, tau=ddays(-1), C_fct="sma_equal")
}
\keyword{internal}

