% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rolling_apply.R
\name{rolling_apply}
\alias{rolling_apply}
\alias{rolling_apply.uts}
\title{Apply Rolling Function}
\usage{
rolling_apply(x, ...)

\method{rolling_apply}{uts}(x, width, FUN, ..., by = NULL, align = "right",
  interior = FALSE)
}
\arguments{
\item{x}{a numeric time series object.}

\item{width}{a finite, non-negative \code{\link[lubridate]{duration}} object, specifying the temporal width of the rolling time window.}

\item{FUN}{a function to be applied to the vector of observation values inside the closed rolling time window.}

\item{by}{a positive \code{\link[lubridate]{duration}} object. If not \code{NULL}, move the rolling time window by steps of this size forward in time, rather than by the observation time differences of \code{x}.}

\item{align}{either \code{"right"}, \code{"left"}, or \code{"center"}. Specifies the alignment of each output time relative to its corresponding time window. Using \code{"right"} gives a causal (i.e. backward-looking) time series operator, while using \code{"left"} gives a purely forward-looking time series operator.}

\item{interior}{logical. Should time windows lie entirely in the interior of the temporal support of \code{x}, i.e. inside the time interval \code{[start(x), end(x)]}?}

\item{\dots}{arguments passed to \code{FUN}.}
}
\description{
Apply a function to the time series values in a closed rolling time window of fixed temporal width.
}
\section{Methods (by class)}{
\itemize{
\item \code{uts}: apply rolling function to \code{"uts"} object.
}}
\examples{
rolling_apply(ex_uts(), width=ddays(1), FUN="mean", by=ddays(0.5))
rolling_apply(ex_uts(), width=ddays(1), FUN="mean")
rolling_apply(ex_uts(), width=ddays(1), FUN="mean", interior=TRUE)
}

