% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_apply.R
\name{rolling_apply}
\alias{rolling_apply}
\alias{rolling_apply.uts}
\title{Apply Rolling Function}
\usage{
rolling_apply(x, ...)

\method{rolling_apply}{uts}(x, width, FUN, ..., by = NULL, align = "right",
  interior = FALSE, use_specialized = TRUE)
}
\arguments{
\item{x}{a numeric time series object.}

\item{width}{a finite, positive \code{\link[lubridate]{duration}} object, specifying the temporal width of the rolling time window.}

\item{FUN}{a function to be applied to the vector of observation values inside the half-open rolling time window.}

\item{by}{a positive \code{\link[lubridate]{duration}} object. If not \code{NULL}, move the rolling time window by steps of this size forward in time, rather than by the observation time differences of \code{x}.}

\item{align}{either \code{"right"}, \code{"left"}, or \code{"center"}. Specifies whether the output times should right- or left-aligned or centered compared to their time window. Using \code{"right"} gives a causal (i.e. backward-looking) time series operator, while using \code{"left"} gives a purely forward-looking time series operator.}

\item{interior}{logical. If \code{TRUE}, then \code{FUN} is only applied if the corresponding time window is in the interior of the temporal support of \code{x}, i.e. inside the time interval \code{[start(x), end(x)]}.}

\item{use_specialized}{logical. Whether to use a fast optimized implementation if available. Currently, the following choices for \code{FUN} are supported: \code{mean}, \code{median}, \code{min}, \code{max}, \code{sum}}

\item{\dots}{arguments passed to \code{FUN}.}
}
\description{
Apply a function to the time series values in a half-open (open on the left, closed on the right) rolling time window of fixed temporal width.
}
\details{
A fast optimized implementation is used automatically for certain choices of \code{FUN}. See the \code{use_specialized} argument for details.
}
\section{Methods (by class)}{
\itemize{
\item \code{uts}: apply rolling function to \code{"uts"} object.
}}
\examples{
# rolling mean, sum, number of observations
rolling_apply(ex_uts(), width=ddays(1), FUN=mean)
rolling_apply(ex_uts(), width=ddays(1), FUN=sum)
rolling_apply(ex_uts(), width=ddays(1), FUN=length)

# move rolling time window by one observation vs. one day at a time
rolling_apply(ex_uts(), width=ddays(1), FUN="mean")
rolling_apply(ex_uts(), width=ddays(1), FUN="mean", by=ddays(1))

# right-align, left-aligned, and centered rolling time window
rolling_apply(ex_uts(), width=ddays(1), FUN=mean)
rolling_apply(ex_uts(), width=ddays(1), FUN=mean, align="left")
rolling_apply(ex_uts(), width=ddays(1), FUN=mean, align="center")

# restrict rolling time window to temporal support of x
rolling_apply(ex_uts(), width=ddays(1), FUN="mean", interior=TRUE)

# specialized vs. general-purpose implementation
rolling_apply(ex_uts(), width=ddays(1), FUN="mean")
rolling_apply(ex_uts(), width=ddays(1), FUN="mean", use_specialized=FALSE)    # same
}

