% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rolling_apply_specialized.R
\name{rolling_apply_specialized}
\alias{rolling_apply_specialized}
\title{Apply Rolling Function (Specialized Implementation)}
\usage{
rolling_apply_specialized(x, ...)
}
\arguments{
\item{x}{a numeric time series object.}

\item{width}{a finite, positive \code{\link[lubridate]{duration}} object, specifying the temporal width of the rolling time window.}
}
\description{
This function provides a fast, specialized implementation of \code{\link{rolling_apply}} for certain choices of \code{FUN} and for \code{by=NULL} (i.e. when moving the rolling time window one observation at a time, rather than by a fixed temporal amount).
}
\details{
It is usually not necessary to call this function. Instead, it is called automatically by \code{\link{rolling_apply}} whenever a specialized implementation is available.
}
\references{
Eckner, A. (2010) \emph{Algorithms for Unevenly Spaced Time Series: Moving Averages and Other Rolling Operators}.
}
\keyword{internal}

