% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rolling_apply_specialized.R
\name{rolling_apply_specialized}
\alias{rolling_apply_specialized}
\alias{rolling_apply_specialized.uts}
\title{Apply Rolling Function (Specialized Implementation)}
\usage{
rolling_apply_specialized(x, ...)

\method{rolling_apply_specialized}{uts}(x, width, FUN, align = "right",
  interior = FALSE, ...)
}
\arguments{
\item{x}{a numeric time series object with finite, non-NA observation values.}

\item{width}{a finite, positive \code{\link[lubridate]{duration}} object, specifying the temporal width of the rolling time window.}

\item{FUN}{a function to be applied to the vector of observation values inside the half-open (open on the left, closed on the right) rolling time window.}

\item{align}{either \code{"right"}, \code{"left"}, or \code{"center"}. Specifies the alignment of each output time relative to its corresponding time window. Using \code{"right"} gives a causal (i.e. backward-looking) time series operator, while using \code{"left"} gives a purely forward-looking time series operator.}

\item{interior}{logical. Should time windows lie entirely in the interior of the temporal support of \code{x}, i.e. inside the time interval \code{[start(x), end(x)]}?}

\item{\ldots}{further arguments passed to or from methods.}
}
\description{
This function provides a fast, specialized implementation of \code{\link{rolling_apply}} for certain choices of \code{FUN} and for \code{by=NULL} (i.e. when moving the rolling time window one observation at a time, rather than by a fixed temporal amount).
}
\details{
It is usually not necessary to call this function, because it is called automatically by \code{\link{rolling_apply}} whenever a specialized implementation is available.
}
\section{Methods (by class)}{
\itemize{
\item \code{uts}: Implementation for \code{"uts"} objects.
}}
\examples{
rolling_apply_specialized(ex_uts(), dhours(12), FUN=length)
rolling_apply_specialized(ex_uts(), dhours(12), FUN=length, align="center")
rolling_apply_specialized(ex_uts(), dhours(12), FUN=length, align="left")

rolling_apply_specialized(ex_uts(), dhours(12), FUN=length)
rolling_apply_specialized(ex_uts(), dhours(12), FUN=length, interior=TRUE)

# Rolling sum
rolling_apply_specialized(ex_uts(), ddays(1), FUN=sum)
rolling_apply_specialized(ex_uts(), ddays(1), FUN=sum) - rolling_apply(ex_uts(), ddays(1), FUN=sum)

# Rolling min/max
rolling_apply_specialized(ex_uts(), ddays(1), FUN=min)
rolling_apply_specialized(ex_uts(), ddays(1), FUN=max)
}
\references{
Eckner, A. (2010) \emph{Algorithms for Unevenly Spaced Time Series: Moving Averages and Other Rolling Operators}.
}
\keyword{internal}

