% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sma.R
\name{sma}
\alias{sma}
\alias{sma.uts}
\title{Simple Moving Average (SMA)}
\usage{
sma(x, ...)

\method{sma}{uts}(x, tau, type = "last", ...)
}
\arguments{
\item{x}{a time series object.}

\item{tau}{a \code{\link[lubridate]{duration}} object, specifying the temporal length of the SMA.}

\item{type}{the type of the SMA. Either \code{"equal"}, \code{"last"}, \code{"next"}, or \code{"linear"}. See below for details}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Calculate a simple moving average (SMA) of a time series.
}
\details{
Currently, four different SMAs types are supported for \code{"uts"} objects. Each type puts different weights on the observation values in the rolling time window of width \code{tau}: \itemize{
\item \code{equal}: Each observation values is weighted equally.
\item \code{last}: Use last-point interpolation. Each observation value is weighted by how long it remained unchanged.
\item \code{next}: Use next-point interpolation. Each observation value is weighted by how long it remained the next (i.e. upcomming) observation.
\item \code{linear}: Use linear interpolation. The behavior is approximately half-way in-between last-point and next-point interpolation.
}
See the reference below for details for precise definitions and on why one would use one SMA type over another.
}
\section{Methods (by class)}{
\itemize{
\item \code{uts}: simple moving average for \code{"uts"} objects.
}}
\examples{
#sma(ex_uts(), ddays(1))
}
\references{
Eckner, A. (2010) \emph{Algorithms for Unevenly Spaced Time Series: Moving Averages and Other Rolling Operators}.
}
\seealso{
\code{\link{ema}} for exponential moving averages.
}

