% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{utsOperators-internal}
\alias{utsOperators-internal}
\title{Internal Functions}
\description{
The internal functions listed below might be of interest to developers seeking to extend the package functionality.
}
\details{
C interfaces:
\itemize{
  \item \code{\link{generic_C_interface}}
}

Helper functions:
\itemize{
  \item \code{\link{check_window_width}}
  \item \code{\link{rolling_apply_specialized}}
  \item \code{\link{rolling_time_window}}
  \item \code{\link{rolling_time_window_indices}}
}

\code{uts} methods:
\itemize{
  \item \code{\link[=rev.uts]{rev}}
  \item \code{\link[=rolling_apply_static]{rolling_apply_static}}
  \item \code{\link{sma_last_R}}
  \item \code{\link{sma_linear_R}}
}
}

